requirements.txt
paho-mqtt
fastapi
uvicorn
jinja2
python-multipart
websockets